% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fram_cvd10_cox.R
\name{fram_cvd10_cox}
\alias{fram_cvd10_cox}
\title{Framingham 10-year Cardiovascular Disease Risk (Cox Model)}
\usage{
fram_cvd10_cox(sex, age, sbp, htn_rx, hdl, tchl, smk, dm)
}
\arguments{
\item{sex}{A character,Sex,Possible values - "Female","Male"}

\item{age}{A number,Age in years,Warning:If age<20years}

\item{sbp}{A number,Systolic Blodd Pressure in mm of Hg}

\item{htn_rx}{A character,Hypertension Treatment Status,Possible values - "Yes","No"}

\item{hdl}{A number,HDL-Cholesterol in mg/dL}

\item{tchl}{A number,Total-Cholesterol in mg/dL}

\item{smk}{A character,Smoking Status,Possible values - "Yes","No"}

\item{dm}{A character,Diabetes Status,Possible values - "Yes","No"}
}
\value{
Framingham 10-year Cardiovascular Disease Risk & Heart/Vascular Age.
}
\description{
This function calculates the 10-year Cardiovascular Disease Risk using Cox Model according to Framingham Heart Study.
}
\details{
This function calculates the 10-year Cardiovascular Disease Risk using Cox Model according to Framingham Heart Study
based on eight predictors provided to it(sex,age,sbp,htn_rx,hdl,tchl,smk,dm).
}
\references{
D'Agostino RB Sr, Vasan RS, Pencina MJ, Wolf PA, Cobain M, Massaro JM, Kannel WB. 
General cardiovascular risk profile for use in primary care: the Framingham Heart Study. 
Circulation. 2008 Feb 12;117(6):743-53. \url{https://doi.org/10.1161/CIRCULATIONAHA.107.699579}.
}
\author{
Biswarup Pramanik
}
